requirements.txt
Adafruit-GPIO==1.0.0
Adafruit-PureIO==0.2.0
Adafruit-SSD1306==1.6.0
AWSIoTPythonSDK==1.0.0
bcdoc==0.12.2
boto3==1.4.0
botocore==1.4.43
chardet==2.3.0
colorama==0.3.2
configparser==3.5.0
docutils==0.12
gpiozero==1.2.0
html5lib==0.999
jmespath==0.9.0
mcpi==0.1.1
numpy==1.8.2
paho-mqtt==1.2
pgzero==1.1
picamera==1.12
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.30
Pillow==2.6.1
pygame==1.9.2a0
Pygments==2.0.1
pygobject==3.14.0
pyserial==2.6
python-apt==0.9.3.12
python-dateutil==2.5.3
python-debian==0.1.27
requests==2.4.3
roman==2.0.0
RPi.GPIO==0.6.2
rsa==3.1.4
RTIMULib==7.2.1
s3transfer==0.1.1
sense-hat==2.2.0
six==1.10.0
smbus==1.1
spidev==3.0
urllib3==1.9.1
